Rename MehrAfzar to MarhAfraz and correct code style based on PSR
