+++ README.md
@@ -1,5 +1,5 @@
 # Laravel SMS Component
-Laravel Package for dealing with Iranian SMS prodivers working with Laravel 5 & 6 
+Laravel Package for dealing with Iranian SMS prodivers working with Laravel 5 & 6
 
 ## Installation
 
@@ -18,7 +18,7 @@ composer require tartan/laravel-iranian-sms
     'IranianSms' => Tartan\IranianSms\Facades\IranianSms::class
 ```
 
-4.Publish the config file 
+4.Publish the config file
 ```bash
 php artisan vendor:publish --provider="Tartan\IranianSms\SmsServiceProvider" --tag=config
 ```
@@ -42,7 +42,7 @@ php artisan vendor:publish --provider="Tartan\IranianSms\SmsServiceProvider" --t
 ### available adapters
 * log (no charge adapter for development purposes)
 * slack (no charge adapter for development and staging environment)
-* mehrafzar مهر افزار
+* mehrafraz مهر افراز
 * kavenegar کاوه نگار
 * smsir اسمس آی آر
 * ghasedak قاصدک
@@ -56,11 +56,11 @@ php artisan vendor:publish --provider="Tartan\IranianSms\SmsServiceProvider" --t
 
 return [
 	//Default SMS gateway
-	'default' => env('IRANIANSMS_DEFAULT','mehrafzar'), 
-	'mehrafzar' => [
-		'gateway' => env('IRANIANSMS_MEHRAFZAR_GATEWAY','http://mehrafraz.com/webservice/Service.asmx?WSDL'),
-		'username' => env('IRANIANSMS_MEHRAFZAR_USERNAME','test'),
-		'password' => env('IRANIANSMS_MEHRAFZAR_PASSWORD','test'),
+	'default' => env('IRANIANSMS_DEFAULT','mehrafraz'),
+	'mehrafraz' => [
+		'gateway' => env('IRANIANSMS_MEHRAFRAZ_GATEWAY','http://mehrafraz.com/webservice/Service.asmx?WSDL'),
+		'username' => env('IRANIANSMS_MEHRAFRAZ_USERNAME','test'),
+		'password' => env('IRANIANSMS_MEHRAFRAZ_PASSWORD','test'),
 	],
 	'kavenegar' => [
 		'gateway' => env('IRANIANSMS_KAVENEGAR_GATEWAY','http://api.kavenegar.com/v1/%s/%s/%s.json/'),
