+++ src/Adapters/AbstractAdapter.php
@@ -4,17 +4,15 @@ namespace Tartan\IranianSms\Adapter;
 
 use Exception;
 
-abstract class AdapterAbstract {
+abstract class AdapterAbstract
+{
+    public function filterNumber(String $number) {
+        $numberCheck = substr($number, -10);
+        if (!preg_match('/9\d{9}$/', $numberCheck)) {
+            throw new Exception('Number format is incorrect');
+        }
 
-	public function filterNumber(String $number) {
+        return $number;
 
-		$numberCheck = substr($number, -10);
-		if (!preg_match('/9\d{9}$/', $numberCheck)) {
-			throw new Exception('Number format is incorrect');
-		}
-
-		return $number;
-
-	}
-
+    }
 }
