+++ src/Adapters/AdapterInterface.php
@@ -2,9 +2,7 @@
 
 namespace Tartan\IranianSms\Adapter;
 
-
-interface AdapterInterface {
-
-	public function send(String $number,String $text);
-
+interface AdapterInterface
+{
+    public function send(String $number,String $text);
 }
