+++ src/Adapters/Exception.php
@@ -1,10 +1,10 @@
 <?php
 
-
 namespace Tartan\IranianSms\Adapters;
+
 use Tartan\IranianSms\Exception as RootException;
 
 class Exception extends RootException
 {
-
+    //
 }
