+++ src/Adapters/Slack.php
