+++ src/Exception.php
@@ -1,10 +1,8 @@
 <?php
 
-
 namespace Tartan\IranianSms;
-
 
 class Exception extends \Exception
 {
-
+    //
 }
