+++ src/Factory.php
@@ -3,7 +3,7 @@
 namespace Tartan\IranianSms;
 
 use Tartan\IranianSms\Adapter\Ghasedak;
-use Tartan\IranianSms\Adapter\MehrAfzar;
+use Tartan\IranianSms\Adapter\MehrAfraz;
 use Tartan\IranianSms\Adapter\KaveNegar;
 use Tartan\IranianSms\Adapter\ParsaSms;
 use Tartan\IranianSms\Adapter\Slack;
@@ -12,21 +12,19 @@ use Tartan\IranianSms\Adapter\SmsLog;
 
 class Factory
 {
-
     function __construct($app)
     {
-
+        //
     }
 
     public function make($adapter = '')
     {
-
         if ($adapter == '') {
             $adapter = config('iranian_sms.default');
         }
         switch ($adapter) {
-            case 'mehrafzar':
-                return new MehrAfzar();
+            case 'mehrafraz':
+                return new MehrAfraz();
                 break;
             case 'kavenegar':
                 return new KaveNegar();
@@ -50,7 +48,6 @@ class Factory
                 throw new Exception('Adapter not defined');
                 break;
         }
-
     }
 }
